refactor(mempool): move test utils to a separate file
