+++ crates/mempool/Cargo.toml
@@ -11,6 +11,8 @@ workspace = true
 [dependencies]
 async-trait.workspace = true
 derive_more.workspace = true
+pretty_assertions = { workspace = true, optional = true }
+starknet-types-core = { workspace = true, optional = true }
 starknet_api.workspace = true
 starknet_mempool_infra.workspace = true
 starknet_mempool_types.workspace = true
@@ -20,8 +22,15 @@ tokio.workspace = true
 assert_matches.workspace = true
 itertools.workspace = true
 mempool_test_utils.workspace = true
-pretty_assertions.workspace = true
 rstest.workspace = true
-starknet-types-core.workspace = true
 starknet_api = { workspace = true, features = ["testing"] }
 tokio.workspace = true
+# Enable test utils feature for integration tests.
+starknet_mempool = { workspace = true, features = ["testing"] }
+
+[features]
+testing = ["pretty_assertions", "starknet-types-core"]
+
+# TODO(Elin): try to remove this once flow test is added.
+[package.metadata.cargo-machete]
+ignored = ["starknet-types-core"]
