+++ crates/mempool/src/lib.rs
@@ -3,3 +3,6 @@ pub mod mempool;
 pub(crate) mod suspended_transaction_pool;
 pub(crate) mod transaction_pool;
 pub(crate) mod transaction_queue;
+
+#[cfg(any(feature = "testing", test))]
+pub mod test_utils;
