+++ crates/mempool/src/mempool_test.rs
@@ -1,5 +1,4 @@
 use std::cmp::Reverse;
-use std::collections::HashMap;
 
 use assert_matches::assert_matches;
 use mempool_test_utils::starknet_api_test_utils::test_resource_bounds_mapping;
@@ -12,15 +11,17 @@ use starknet_api::test_utils::invoke::executable_invoke_tx;
 use starknet_api::transaction::{Tip, TransactionHash, ValidResourceBounds};
 use starknet_api::{contract_address, felt, invoke_tx_args, nonce, patricia_key};
 use starknet_mempool_types::errors::MempoolError;
-use starknet_mempool_types::mempool_types::{AccountState, CommitBlockArgs};
+use starknet_mempool_types::mempool_types::AccountState;
 
 use crate::mempool::{AccountToNonce, AddTransactionArgs, Mempool, TransactionReference};
+use crate::test_utils::{add_tx, add_tx_expect_error, commit_block, get_txs_and_assert_expected};
 use crate::transaction_pool::TransactionPool;
 use crate::transaction_queue::transaction_queue_test_utils::{
     TransactionQueueContent,
     TransactionQueueContentBuilder,
 };
 use crate::transaction_queue::TransactionQueue;
+use crate::{add_tx_input, tx};
 
 // Utils.
 
@@ -151,102 +152,6 @@ impl FromIterator<TransactionReference> for TransactionQueue {
         }
         queue
     }
-}
-
-#[track_caller]
-fn add_tx(mempool: &mut Mempool, input: &AddTransactionArgs) {
-    assert_eq!(mempool.add_tx(input.clone()), Ok(()));
-}
-
-#[track_caller]
-fn add_tx_expect_error(
-    mempool: &mut Mempool,
-    input: &AddTransactionArgs,
-    expected_error: MempoolError,
-) {
-    assert_eq!(mempool.add_tx(input.clone()), Err(expected_error));
-}
-
-#[track_caller]
-fn commit_block(mempool: &mut Mempool, nonces: impl IntoIterator<Item = (&'static str, u8)>) {
-    let nonces = HashMap::from_iter(
-        nonces.into_iter().map(|(address, nonce)| (contract_address!(address), nonce!(nonce))),
-    );
-    let args = CommitBlockArgs { nonces };
-
-    assert_eq!(mempool.commit_block(args), Ok(()));
-}
-
-#[track_caller]
-fn get_txs_and_assert_expected(mempool: &mut Mempool, n_txs: usize, expected_txs: &[Transaction]) {
-    let txs = mempool.get_txs(n_txs).unwrap();
-    assert_eq!(txs, expected_txs);
-}
-
-/// Creates an executable invoke transaction with the given field subset (the rest receive default
-/// values).
-macro_rules! tx {
-    (tip: $tip:expr, tx_hash: $tx_hash:expr, sender_address: $sender_address:expr,
-        tx_nonce: $tx_nonce:expr, resource_bounds: $resource_bounds:expr) => {{
-            let sender_address = contract_address!($sender_address);
-            Transaction::Invoke(executable_invoke_tx(invoke_tx_args!{
-                sender_address: sender_address,
-                tx_hash: TransactionHash(StarkHash::from($tx_hash)),
-                tip: Tip($tip),
-                nonce: nonce!($tx_nonce),
-                resource_bounds: $resource_bounds,
-            }))
-    }};
-    (tip: $tip:expr, tx_hash: $tx_hash:expr, sender_address: $sender_address:expr, tx_nonce: $tx_nonce:expr) => {
-        tx!(tip: $tip, tx_hash: $tx_hash, sender_address: $sender_address, tx_nonce: $tx_nonce, resource_bounds: ValidResourceBounds::AllResources(test_resource_bounds_mapping()))
-    };
-    (tx_hash: $tx_hash:expr, sender_address: $sender_address:expr, tx_nonce: $tx_nonce:expr) => {
-        tx!(tip: 0, tx_hash: $tx_hash, sender_address: $sender_address, tx_nonce: $tx_nonce)
-    };
-    (tip: $tip:expr, tx_hash: $tx_hash:expr, sender_address: $sender_address:expr) => {
-        tx!(tip: $tip, tx_hash: $tx_hash, sender_address: $sender_address, tx_nonce: 0)
-    };
-    (tx_hash: $tx_hash:expr, tx_nonce: $tx_nonce:expr) => {
-        tx!(tip: 0, tx_hash: $tx_hash, sender_address: "0x0", tx_nonce: $tx_nonce)
-    };
-    (tx_nonce: $tx_nonce:expr) => {
-        tx!(tip: 0, tx_hash: 0, sender_address: "0x0", tx_nonce: $tx_nonce)
-    };
-}
-
-/// Creates an input for `add_tx` with the given field subset (the rest receive default values).
-macro_rules! add_tx_input {
-    (tip: $tip:expr, tx_hash: $tx_hash:expr, sender_address: $sender_address:expr,
-        tx_nonce: $tx_nonce:expr, account_nonce: $account_nonce:expr, resource_bounds: $resource_bounds:expr) => {{
-        let tx = tx!(tip: $tip, tx_hash: $tx_hash, sender_address: $sender_address, tx_nonce: $tx_nonce, resource_bounds: $resource_bounds);
-        let address = contract_address!($sender_address);
-        let account_nonce = nonce!($account_nonce);
-        let account_state = AccountState { address, nonce: account_nonce};
-
-        AddTransactionArgs { tx, account_state }
-    }};
-    (tip: $tip:expr, tx_hash: $tx_hash:expr, sender_address: $sender_address:expr,
-        tx_nonce: $tx_nonce:expr, account_nonce: $account_nonce:expr) => {{
-            add_tx_input!(tip: $tip, tx_hash: $tx_hash, sender_address: $sender_address, tx_nonce: $tx_nonce, account_nonce: $account_nonce, resource_bounds: ValidResourceBounds::AllResources(test_resource_bounds_mapping()))
-    }};
-    (tx_hash: $tx_hash:expr, sender_address: $sender_address:expr, tx_nonce: $tx_nonce:expr, account_nonce: $account_nonce:expr) => {
-        add_tx_input!(tip: 0, tx_hash: $tx_hash, sender_address: $sender_address, tx_nonce: $tx_nonce, account_nonce: $account_nonce)
-    };
-    (tip: $tip:expr, tx_hash: $tx_hash:expr, sender_address: $sender_address:expr) => {
-        add_tx_input!(tip: $tip, tx_hash: $tx_hash, sender_address: $sender_address, tx_nonce: 0, account_nonce: 0)
-    };
-    (tx_hash: $tx_hash:expr, tx_nonce: $tx_nonce:expr, account_nonce: $account_nonce:expr) => {
-        add_tx_input!(tip: 1, tx_hash: $tx_hash, sender_address: "0x0", tx_nonce: $tx_nonce, account_nonce: $account_nonce)
-    };
-    (tx_nonce: $tx_nonce:expr, account_nonce: $account_nonce:expr) => {
-        add_tx_input!(tip: 1, tx_hash: 0, sender_address: "0x0", tx_nonce: $tx_nonce, account_nonce: $account_nonce)
-    };
-    (tip: $tip:expr, tx_hash: $tx_hash:expr) => {
-        add_tx_input!(tip: $tip, tx_hash: $tx_hash, sender_address: "0x0", tx_nonce: 0, account_nonce: 0)
-    };
-    (tx_hash: $tx_hash:expr, tx_nonce: $tx_nonce:expr) => {
-        add_tx_input!(tip: 0, tx_hash: $tx_hash, sender_address: "0x0", tx_nonce: $tx_nonce, account_nonce: 0)
-    };
 }
 
 // Fixtures.
